requirements.txt
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
yfinance>=0.2.31
pyyaml>=6.0.1
